+++ TencentCloud/Cdb/V20170320/Models/DeleteAuditRuleTemplatesRequest.cs
@@ -25,7 +25,7 @@ namespace TencentCloud.Cdb.V20170320.Models
     {
         
         /// <summary>
-        /// 审计规则模版ID。
+        /// 审计规则模板ID。
         /// </summary>
         [JsonProperty("RuleTemplateIds")]
         public string[] RuleTemplateIds{ get; set; }