+++ TencentCloud/Cdwdoris/V20211228/Models/InstanceInfo.cs
@@ -277,6 +277,7 @@ namespace TencentCloud.Cdwdoris.V20211228.Models
 
         /// <summary>
         /// 组件信息
+        /// 注：这里返回类型实际为map[string]struct类型，并非显示的string类型，可以参考“示例值”进行数据的解析。
         /// 注意：此字段可能返回 null，表示取不到有效值。
         /// </summary>
         [JsonProperty("Components")]