+++ TencentCloud/Essbasic/V20210526/Models/ChannelCancelMultiFlowSignQRCodeRequest.cs
@@ -39,7 +39,7 @@ namespace TencentCloud.Essbasic.V20210526.Models
         public Agent Agent{ get; set; }
 
         /// <summary>
-        /// 需要取消签署的二维码ID，为32位字符串。由[创建一码多扫流程签署二维码](https://qian.tencent.com/developers/partnerApis/templates/ChannelCreateMultiFlowSignQRCode)返回
+        /// 需要取消的签署码ID，为32位字符串。由[创建一码多签签署码](https://qian.tencent.com/developers/partnerApis/templates/ChannelCreateMultiFlowSignQRCode)返回
         /// </summary>
         [JsonProperty("QrCodeId")]
         public string QrCodeId{ get; set; }