+++ TencentCloud/Sqlserver/V20180328/Models/DescribeInstanceTradeParameterRequest.cs
@@ -55,7 +55,7 @@ namespace TencentCloud.Sqlserver.V20180328.Models
         public string InstanceType{ get; set; }
 
         /// <summary>
-        /// 购买实例的宿主机磁盘类型,CLOUD_HSSD-虚拟机加强型SSD云盘，CLOUD_TSSD-虚拟机极速型SSD云盘，CLOUD_BSSD-虚拟机通用型SSD云盘
+        /// 购买实例的宿主机磁盘类型,CLOUD_HSSD-云服务器加强型SSD云盘，CLOUD_TSSD-云服务器极速型SSD云盘，CLOUD_BSSD-云服务器通用型SSD云盘
         /// </summary>
         [JsonProperty("MachineType")]
         public string MachineType{ get; set; }