+++ TencentCloud/Tcss/V20201101/Models/DescribePurchaseStateInfoResponse.cs
@@ -102,6 +102,13 @@ namespace TencentCloud.Tcss.V20201101.Models
         public string SubState{ get; set; }
 
         /// <summary>
+        /// 计费key
+        /// 注意：此字段可能返回 null，表示取不到有效值。
+        /// </summary>
+        [JsonProperty("InquireKey")]
+        public string InquireKey{ get; set; }
+
+        /// <summary>
         /// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
         /// </summary>
         [JsonProperty("RequestId")]
@@ -124,6 +131,7 @@ namespace TencentCloud.Tcss.V20201101.Models
             this.SetParamSimple(map, prefix + "GivenAuthorizedCnt", this.GivenAuthorizedCnt);
             this.SetParamSimple(map, prefix + "BeginTime", this.BeginTime);
             this.SetParamSimple(map, prefix + "SubState", this.SubState);
+            this.SetParamSimple(map, prefix + "InquireKey", this.InquireKey);
             this.SetParamSimple(map, prefix + "RequestId", this.RequestId);
         }
     }