+++ CHANGELOG.md
@@ -1,3 +1,166 @@
+# Release 3.0.929
+
+## 云防火墙(cfw) 版本：2019-09-04
+
+### 第 52 次发布
+
+发布时间：2024-01-05 01:09:38
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+<font color="#dd0000">**删除接口**：</font>
+
+* ModifyPublicIPSwitchStatus
+
+
+
+## 主机安全(cwp) 版本：2018-02-28
+
+### 第 106 次发布
+
+发布时间：2024-01-05 01:11:37
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [BruteAttackInfo](https://cloud.tencent.com/document/api/296/19867#BruteAttackInfo)
+
+	* 新增成员：DataFrom
+
+
+
+
+## T-Sec-堡垒机（BH）(dasb) 版本：2019-10-18
+
+### 第 21 次发布
+
+发布时间：2024-01-05 01:13:10
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [Command](https://cloud.tencent.com/document/api/1025/74416#Command)
+
+	* 新增成员：Size
+
+* [SearchCommandResult](https://cloud.tencent.com/document/api/1025/74416#SearchCommandResult)
+
+	* 新增成员：Size
+
+
+
+
+## 腾讯混元大模型(hunyuan) 版本：2023-09-01
+
+### 第 2 次发布
+
+发布时间：2024-01-05 01:17:08
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [GetEmbedding](https://cloud.tencent.com/document/api/1729/102832)
+
+新增数据结构：
+
+* [EmbeddingData](https://cloud.tencent.com/document/api/1729/101838#EmbeddingData)
+* [EmbeddingUsage](https://cloud.tencent.com/document/api/1729/101838#EmbeddingUsage)
+
+
+
+## 文字识别(ocr) 版本：2018-11-19
+
+### 第 126 次发布
+
+发布时间：2024-01-05 01:21:40
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [SmartStructuralOCRV2](https://cloud.tencent.com/document/api/866/90013)
+
+	* 新增入参：EnableSealRecognize
+
+
+
+
+## 腾讯健康组学平台(omics) 版本：2022-11-28
+
+### 第 11 次发布
+
+发布时间：2024-01-05 01:21:57
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [RunApplication](https://cloud.tencent.com/document/api/1643/89094)
+
+	* 新增入参：NFOption
+
+	* <font color="#dd0000">**修改入参**：</font>ProjectId, Option
+
+
+
+
+## 容器安全服务(tcss) 版本：2020-11-01
+
+### 第 59 次发布
+
+发布时间：2024-01-05 01:25:17
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DescribePurchaseStateInfo](https://cloud.tencent.com/document/api/1285/65466)
+
+	* 新增出参：InquireKey
+
+
+
+
+## 微服务平台 TSF(tsf) 版本：2018-03-26
+
+### 第 103 次发布
+
+发布时间：2024-01-05 01:28:46
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增数据结构：
+
+* [BusinessLogConfigAssociatedGroup](https://cloud.tencent.com/document/api/649/36099#BusinessLogConfigAssociatedGroup)
+
+修改数据结构：
+
+* [BusinessLogConfig](https://cloud.tencent.com/document/api/649/36099#BusinessLogConfig)
+
+	* 新增成员：ConfigAssociatedGroupList
+
+
+
+
 # Release 3.0.928
 
 ## 云数据仓库 ClickHouse(cdwch) 版本：2020-09-15